QmlObject: fix comparison of mixed-signedness values - idiomatically
In general, the fix for mixed-signedness comparisons is to lift the
comparison into unsigned, not signed, space. Usually, that means that
you need an additional < 0 check for the signed values, but here we
know that QString::size() is always >= 0, so it suffices to just cast
size() to size_t, the unsigned version of qsizetype.

(cherry picked from commit d8e18d355a681fb1213b6b3f28df1ccc03a91291)

+++ src/qmldom/qqmldomelements.cpp
@@ -772,7 +772,7 @@ void QmlObject::writeOut(DomItem &self, OutWriter &ow, QString onTarget) const
             // check for an empty line before the current element, and preserve it
             int preNewlines = 0;
             quint32 start = el.first.offset;
-            if (start != posOfNewElements && code.size() >= start) {
+            if (start != posOfNewElements && size_t(code.size()) >= start) {
                 while (start != 0) {
                     QChar c = code.at(--start);
                     if (c == u'\n') {
